+++ src/algorithms/mod.rs
@@ -1,5 +1,7 @@
+mod field_of_movement;
 mod fov;
 mod pathfinding;
 
+pub use field_of_movement::field_of_movement;
 pub use fov::{directional_fov, range_fov};
 pub use pathfinding::a_star;